requirements.txt
flask==3.0.3
PyMuPDF==1.24.9
openai==0.28.1
python-dotenv==1.0.1

